% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{character_labels}
\alias{character_labels}
\title{A table with localized character labels and descriptors}
\format{
A tibble with 173,013 rows and 7 variables:
\describe{
\item{locale}{The locale name.}
\item{character_label_patterns}{character_label_patterns}
\item{character_labels}{character_labels}
}
}
\usage{
character_labels
}
\description{
The \code{character_labels} table contains localized data for character labels
across 571 locales.
}
\keyword{datasets}
