% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{characters}
\alias{characters}
\title{A table with localized character data}
\format{
A tibble with 173,013 rows and 7 variables:
\describe{
\item{locale}{The locale name.}
\item{exemplar_characters}{exemplar_characters}
\item{auxiliary}{auxiliary}
\item{index}{index}
\item{numbers}{numbers}
\item{punctuation}{punctuation}
\item{ellipsis}{ellipsis}
\item{leninent_scope_general}{leninent_scope_general}
\item{leninent_scope_date}{leninent_scope_date}
\item{leninent_scope_number}{leninent_scope_number}
\item{stricter_scope_number}{stricter_scope_number}
}
}
\usage{
characters
}
\description{
The \code{characters} table contains localized character data across 571 locales.
}
\keyword{datasets}
