% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\docType{data}
\name{characters_elements}
\alias{characters_elements}
\title{A list of elements from the \code{characters} table}
\format{
An object of class \code{list} of length 12.
}
\usage{
characters_elements
}
\value{
A named list.
}
\description{
A list of elements from the \code{characters} table
}
\keyword{datasets}
