% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\name{cldr_character_labels}
\alias{cldr_character_labels}
\title{Get localized values from the \code{character_labels} dataset}
\usage{
cldr_character_labels(locale = "en", param = c("patterns", "labels"))
}
\value{
A named list.
}
\description{
The \code{character_labels} table contains information on character patterns
and character labels across 571 locales. The \code{cldr_character_labels()}
function allows one to extract parameter values from the table by
supplying the \code{locale} and one of the following parameter names:
\itemize{
\item \code{"patterns"}: corresponds to the \code{character_label_patterns} column in
\code{character_labels}
\item \code{"labels"}: is the \code{character_labels} column in \code{character_labels}
}
}
