% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\name{cldr_currency}
\alias{cldr_currency}
\title{Get a single localized value from the \code{currencies} dataset}
\usage{
cldr_currency(
  locale = "en",
  currency = currency_codes$USD,
  element = currencies_elements$currency_symbol
)
}
\arguments{
\item{locale}{The locale ID for which to obtain the data from the
\code{currencies} table.}

\item{currency}{The currency code (e.g., \code{"USD"}, \code{"EUR"}, etc.). A valid set
of currency codes can be accessed through the \code{currency_codes} list object.}

\item{element}{The element from which information will be obtained for the
specified \code{locale}. A valid set of currency element can be accessed through
the \code{currencies_elements} list object.}
}
\value{
A length one character vector.
}
\description{
The \code{currencies} table contains information of currency codes and localized
display names and symbols across 571 locales. The \code{cldr_currency()} function
allows one to extract a single element value from the table by supplying
the \code{locale}, the currency code (\code{currency}), and one of the following
\code{element} names:
\itemize{
\item \code{"currency_symbol"}
\item \code{"currency_symbol_narrow"}
\item \code{"currency_display_name"}
\item \code{"currency_display_name_count_1"}
\item \code{"currency_display_name_count_other"}
}
}
