% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{currencies}
\alias{currencies}
\title{A table with localized currency attributes and descriptors}
\format{
A tibble with 173,013 rows and 7 variables:
\describe{
\item{locale}{The locale name.}
\item{currency_code}{The three-letter code that identifies the currency.}
\item{currency_display_name}{currency_display_name}
\item{currency_symbol}{currency_symbol}
\item{currency_symbol_narrow}{currency_symbol_narrow}
\item{currency_display_name_count_1}{currency_display_name_count_1}
\item{currency_display_name_count_other}{currency_display_name_count_other}
}
}
\usage{
currencies
}
\description{
The \code{currencies} table contains localized data for currency-related entities
across 571 locales.
}
\keyword{datasets}
