% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\docType{data}
\name{currencies_elements}
\alias{currencies_elements}
\title{A list of elements from the \code{currencies} table}
\format{
An object of class \code{list} of length 5.
}
\usage{
currencies_elements
}
\value{
A named list.
}
\description{
A list of elements from the \code{currencies} table
}
\keyword{datasets}
