% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\docType{data}
\name{currency_codes}
\alias{currency_codes}
\title{A list of currency codes for use in the \code{cldr_currency()} function}
\format{
An object of class \code{list} of length 303.
}
\usage{
currency_codes
}
\value{
A named list.
}
\description{
A list of currency codes for use in the \code{cldr_currency()} function
}
\keyword{datasets}
