% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dates}
\alias{dates}
\title{A table with localized date attributes and descriptors}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 571 rows and 29 columns.
}
\usage{
dates
}
\description{
The \code{dates} table contains localized data for constructing dates and times
across 571 locales. There are 571
rows and the following 29 columns:
}
\details{
\code{locale} (\code{character})
\code{months_format_abbrev} (\verb{named list [length of 12]})
\code{months_format_narrow} (\verb{named list [length of 12]})
\code{months_format_wide} (\verb{named list [length of 12]})
\code{days_standalone_narrow} (\verb{named list [length of 7]})
\code{days_standalone_short} (\verb{named list [length of 7]})
\code{days_standalone_wide} (\verb{named list [length of 7]})
\code{quarters_format_abbrev} (\verb{named list [length of 4]})
\code{quarters_format_narrow} (\verb{named list [length of 4]})
\code{quarters_format_wide} (\verb{named list [length of 4]})
\code{quarters_standalone_abbrev} (\verb{named list [length of 4]})
\code{quarters_standalone_narrow} (\verb{named list [length of 4]})
\code{quarters_standalone_wide} (\verb{named list [length of 4]})
\code{dayperiods_format_abbrev} (\verb{named list [variable length]})
\code{dayperiods_format_narrow} (\verb{named list [variable length]})
\code{dayperiods_format_wide} (\verb{named list [variable length]})
\code{dayperiods_standalone_abbrev} (\verb{named list [variable length]})
\code{dayperiods_standalone_narrow} (\verb{named list [variable length]})
\code{dayperiods_standalone_wide} (\verb{named list [variable length]})
\code{eras_abbrev} (\verb{named list [length of 4]})
\code{eras_names} (\verb{named list [length of 4]})
\code{eras_narrow} (\verb{named list [length of 4]})
\code{date_formats} (\verb{named list [variable length]})
\code{date_skeletons} (\verb{named list [length of 4]})
\code{time_formats} (\verb{named list [variable length]})
\code{time_skeletons} (\verb{named list [variable length]})
\code{date_time_available_formats} (\verb{named list [variable length]})
\code{date_time_append_items} (\verb{named list [length of 11]})
\code{date_time_interval_formats} (\verb{named list [variable length]})
}
\keyword{datasets}
