% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\docType{data}
\name{dates_elements}
\alias{dates_elements}
\title{A list of elements from the \code{dates} table}
\format{
An object of class \code{list} of length 28.
}
\usage{
dates_elements
}
\value{
A named list.
}
\description{
A list of elements from the \code{dates} table
}
\keyword{datasets}
