% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dates_generic}
\alias{dates_generic}
\title{A table with localized generic date attributes and descriptors}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 574 rows and 38 columns.
}
\usage{
dates_generic
}
\description{
The \code{dates_generic} table contains localized data for constructing dates and
times across 574 locales. There are
574 rows and the following 38
columns:
\itemize{
\item \code{locale} (\code{character})
\item \code{months_format_abbrev} (\verb{named list [length of 12]})
\item \code{months_format_narrow} (\verb{named list [length of 12]})
\item \code{months_format_wide} (\verb{named list [length of 12]})
\item \code{months_standalone_abbrev} (\verb{named list [length of 12]})
\item \code{months_standalone_narrow} (\verb{named list [length of 12]})
\item \code{months_standalone_wide} (\verb{named list [length of 12]})
\item \code{days_format_abbrev} (\verb{named list [length of 7]})
\item \code{days_format_narrow} (\verb{named list [length of 7]})
\item \code{days_format_short} (\verb{named list [length of 7]})
\item \code{days_format_wide} (\verb{named list [length of 7]})
\item \code{days_standalone_abbrev} (\verb{named list [length of 7]})
\item \code{days_standalone_narrow} (\verb{named list [length of 7]})
\item \code{days_standalone_short} (\verb{named list [length of 7]})
\item \code{days_standalone_wide} (\verb{named list [length of 7]})
\item \code{quarters_format_abbrev} (\verb{named list [length of 4]})
\item \code{quarters_format_narrow} (\verb{named list [length of 4]})
\item \code{quarters_format_wide} (\verb{named list [length of 4]})
\item \code{quarters_standalone_abbrev} (\verb{named list [length of 4]})
\item \code{quarters_standalone_narrow} (\verb{named list [length of 4]})
\item \code{quarters_standalone_wide} (\verb{named list [length of 4]})
\item \code{dayperiods_format_abbrev} (\verb{named list [variable length]})
\item \code{dayperiods_format_narrow} (\verb{named list [variable length]})
\item \code{dayperiods_format_wide} (\verb{named list [variable length]})
\item \code{dayperiods_standalone_abbrev} (\verb{named list [variable length]})
\item \code{dayperiods_standalone_narrow} (\verb{named list [variable length]})
\item \code{dayperiods_standalone_wide} (\verb{named list [variable length]})
\item \code{eras_abbrev} (\verb{named list [length of 4]})
\item \code{eras_names} (\verb{named list [length of 4]})
\item \code{eras_narrow} (\verb{named list [length of 4]})
\item \code{date_formats} (\verb{named list [variable length]})
\item \code{date_skeletons} (\verb{named list [length of 4]})
\item \code{time_formats} (\verb{named list [variable length]})
\item \code{time_skeletons} (\verb{named list [variable length]})
\item \code{date_time_patterns} (\verb{named list [length of 4]})
\item \code{date_time_available_formats} (\verb{named list [variable length]})
\item \code{date_time_append_items} (\verb{named list [length of 11]})
\item \code{date_time_interval_formats} (\verb{named list [variable length]})
}
}
\keyword{datasets}
