% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{default_locales}
\alias{default_locales}
\title{A table containing a mapping of default locale names to base locales}
\format{
A tibble with 226 rows and 2 variables:
\describe{
\item{default_locale}{The expanded locale names (e.g., \code{"en-US"}) that do not
normally appear within the \strong{cldr} datasets but are valid aliases for base
locale names (e.g., \code{"en"}).}
\item{base_locale}{The corresponding base name for the locale. This omits the
region particle from the locale name.}
}
}
\usage{
default_locales
}
\description{
This is a table that contains base locale names (e.g., \code{"en"}, \code{"de"})
alongside their default locale names. This indicates that \code{"en"} maps to
\code{"en-US"} and \code{"de"} should map to \code{"de-DE"}. Throughout the \strong{cldr}
datasets, base names are used instead of their expanded equivalents.
}
\keyword{datasets}
