% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\name{get_currency_value}
\alias{get_currency_value}
\title{Get a single localized value from the \code{currencies} dataset}
\usage{
get_currency_value(
  locale = "en",
  currency = "USD",
  param = c("sym", "sym_narrow", "displ_name", "displ_ct_1", "displ_ct_oth")
)
}
\value{
A length one character vector.
}
\description{
The \code{currencies} table contains information of currency codes and localized
display names and symbols across 571 locales. The \code{get_currency_value()}
function allows one to extract a single parameter value from the table by
supplying the \code{locale}, the currency code, and one of the following parameter
names.
\itemize{
\item \code{"sym"}: corresponds to the \code{currency_symbol} column in \code{currencies}
\item \code{"sym_narrow"}: is the \code{currency_symbol_narrow} column in \code{currencies}
\item \code{"displ_name"}: is \code{currency_display_name}
\item \code{"displ_ct_1"}: is \code{currency_display_name_count_1}
\item \code{"displ_ct_oth"}: is \code{currency_display_name_count_other}
}
}
