% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{layout}
\alias{layout}
\title{A table with localized layout data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 574 rows and 3 columns.
}
\usage{
layout
}
\description{
The \code{layout} table contains data on text layout across
574 locales. There are 574 rows and
the following 3 columns:
\itemize{
\item \code{locale} (\code{character})
\item \code{character_order} (\code{character})
\item \code{line_order} (\code{character})
}
}
\keyword{datasets}
