% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\docType{data}
\name{locale_elements}
\alias{locale_elements}
\title{A list of elements from the \code{locale_names} table}
\format{
An object of class \code{list} of length 4.
}
\usage{
locale_elements
}
\value{
A named list.
}
\description{
A list of elements from the \code{locale_names} table
}
\keyword{datasets}
