% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{locale_names}
\alias{locale_names}
\title{A table with localized language, script, and territory names}
\format{
A tibble with 571 rows and 5 variables:
\describe{
\item{locale}{The locale name.}
\item{lang_names}{A list column containing named lists for all localized
language names.}
\item{script_names}{A list column containing named lists for all localized
script names.}
\item{territory_names}{A list column containing named lists for all localized
territory names.}
\item{variant_names}{A list column containing named lists for all localized
variant names.}
}
}
\usage{
locale_names
}
\description{
The \code{locale_names} table contains localized names for languages
(\code{lang_names}), script names (\code{script_names}), names for territories
(\code{territory_names}), and names for variants (\code{variant_names}).
}
\keyword{datasets}
