% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{num_system_digits}
\alias{num_system_digits}
\title{Vectors of digits from various numbering systems}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 53 rows and 2 columns.
}
\usage{
num_system_digits
}
\description{
The \code{num_system_digits} table contains vectors of digits for different
numbering systems. Each vector contains 10 elements (comprising the numbers
\code{0} to \code{9}) and there are are 53 rows in total. The
following columns are included:
\itemize{
\item \code{script} (\code{character})
\item \code{digits} (\verb{list [length of 10]})
}
}
\keyword{datasets}
