% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{numbers}
\alias{numbers}
\title{A table with localized numerical attributes and descriptors}
\format{
A tibble with 571 rows and 26 variables:
\describe{
\item{locale}{The locale name.}
\item{default_numbering_system}{default_numbering_system}
\item{other_numbering_systems}{other_numbering_systems}
\item{minimum_grouping_digits}{minimum_grouping_digits}
\item{decimal}{decimal}
\item{group}{group}
\item{list}{list}
\item{percent_sign}{percent_sign}
\item{plus_sign}{plus_sign}
\item{minus_sign}{minus_sign}
\item{approx_sign}{approx_sign}
\item{exp_sign}{exp_sign}
\item{sup_exp}{sup_exp}
\item{per_mille}{per_mille}
\item{infinity}{infinity}
\item{nan}{nan}
\item{time_sep}{time_sep}
\item{approx_pattern}{approx_pattern}
\item{at_least_pattern}{at_least_pattern}
\item{at_most_pattern}{at_most_pattern}
\item{range_pattern}{range_pattern}
\item{decimal_format}{decimal_format}
\item{sci_format}{sci_format}
\item{percent_format}{percent_format}
\item{currency_format}{currency_format}
\item{accounting_format}{accounting_format}
}
}
\usage{
numbers
}
\description{
The \code{numbers} table contains localized data for number-related entities
across 571 locales.
}
\keyword{datasets}
