% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\docType{data}
\name{numbers_elements}
\alias{numbers_elements}
\title{A list of elements from the \code{numbers} table}
\format{
An object of class \code{list} of length 26.
}
\usage{
numbers_elements
}
\value{
A named list.
}
\description{
A list of elements from the \code{numbers} table
}
\keyword{datasets}
