% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{script_metadata}
\alias{script_metadata}
\title{A table with metadata for a wide variety of script types}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 170 rows and 11 columns.
}
\usage{
script_metadata
}
\description{
The \code{script_metadata} table contains metadata for various script types.
There are 170 rows and the following 11 columns:
\itemize{
\item \code{script} (\code{character})
\item \code{sample_char} (\code{character})
\item \code{rank} (\code{integer})
\item \code{script} (\code{character})
\item \code{rtl} (\code{character})
\item \code{lb_letters} (\code{character})
\item \code{has_case} (\code{character})
\item \code{shaping_req} (\code{character})
\item \code{ime} (\code{character})
\item \code{density} (\code{integer})
\item \code{origin_country} (\code{character})
\item \code{likely_lang} (\code{character})
}
}
\keyword{datasets}
