% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{start_of_week}
\alias{start_of_week}
\title{A table with the starting day of the week across territories}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 151 rows and 2 columns.
}
\usage{
start_of_week
}
\description{
The \code{start_of_week} table contains the day names for the start of the week
(e.g., \code{"sun"}, \code{"mon"}, etc.) for a given territory (which is typically a
2-letter country code). The following columns are included:
\itemize{
\item \code{territory} (\code{character})
\item \code{day_of_week} (\code{character})
}
}
\keyword{datasets}
