% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tz_bcp_id}
\alias{tz_bcp_id}
\title{A table with BCP47 Olson/IANA-style and canonical time zone IDs}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 593 rows and 3 columns.
}
\usage{
tz_bcp_id
}
\description{
The \code{tz_bcp_id} table provides a lookup for converting between BCP47
Olson/IANA-style time zone IDs and the canonical forms (according to BCP47).
There are 593 rows and the following 3 columns:
\itemize{
\item \code{tz_bcp_id} (\code{character})
\item \code{tz_canonical} (\code{character})
\item \code{description} (\code{character})
}
}
\keyword{datasets}
