% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tz_formats}
\alias{tz_formats}
\title{A table with localized time zone formatting information}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 574 rows and 8 columns.
}
\usage{
tz_formats
}
\description{
The \code{tz_formats} table contains localized formatting information across all
locales. There are 574 rows and the following 7 columns:
\itemize{
\item \code{locale} (\code{character})
\item \code{hour_format} (\code{character})
\item \code{gmt_format} (\code{character})
\item \code{gmt_zero_format} (\code{character})
\item \code{region_format} (\code{character})
\item \code{region_format_daylight} (\code{character})
\item \code{region_format_standard} (\code{character})
\item \code{region_format_fallback} (\code{character})
}
}
\keyword{datasets}
