% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tz_metazone_users}
\alias{tz_metazone_users}
\title{A table that links canonical tz names with their metazone}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 293 rows and 4 columns.
}
\usage{
tz_metazone_users
}
\description{
The \code{tz_metazone_users} table allows for a lookup of canonical time zone
name to which metazone each uses. As an example, the canonical time zone
\code{America/Vancouver} corresponds to the \code{America_Pacific} metazone (this is
the long ID, but there is often a short ID available as well). The
\code{metazone_long_id} can be used to get a localized metazone name by use of
the \link{tz_metazone_names} table.

There are 293 rows and the following 4 columns:
\itemize{
\item \code{canonical_tz_name} (\code{character})
\item \code{territory} (\code{character})
\item \code{metazone_long_id} (\code{character})
\item \code{metazone_short_id} (\code{character})
}
}
\keyword{datasets}
