% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tz_names}
\alias{tz_names}
\title{A table with names for time zone exemplar cities}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 571 rows and 433 columns.
}
\usage{
tz_names
}
\description{
The \code{tz_names} table contains localized names for all exemplar cities used in
time zone names. There are 571 rows and a column for each
exemplar city name (comprising 432 columns).
}
\keyword{datasets}
