% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{all_currency_codes}
\alias{all_currency_codes}
\title{A vector containing every currency code}
\format{
An object of class \code{character} of length 303.
}
\usage{
all_currency_codes
}
\description{
This is a vector of the 303 currency codes that
are used in the \link{currencies} dataset within the \strong{i18n} package.
}
\keyword{datasets}
