% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\name{cldr_character_labels}
\alias{cldr_character_labels}
\title{Get localized values from the \link{character_labels} dataset}
\usage{
cldr_character_labels(locale = "en", element = c("patterns", "labels"))
}
\arguments{
\item{locale}{The locale ID for which to obtain the data from the
\code{character_labels} table.}

\item{element}{The element from which information will be obtained for the
specified \code{locale}.}
}
\value{
A named list.
}
\description{
The \link{character_labels} table contains information on character patterns
and character labels across 571 locales. The
\code{cldr_character_labels()} function allows one to extract element values from
the table by supplying the \code{locale} and one of the following element names:
\itemize{
\item \code{"patterns"}: corresponds to the \code{character_label_patterns} column in
\link{character_labels}
\item \code{"labels"}: is the \code{character_labels} column in \link{character_labels}
}
}
