% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\name{cldr_currency}
\alias{cldr_currency}
\title{Get a single localized value from the \code{currencies} dataset}
\usage{
cldr_currency(
  locale = "en",
  currency = "USD",
  param = c("sym", "sym_narrow", "displ_name", "displ_ct_1", "displ_ct_oth")
)
}
\arguments{
\item{locale}{The locale ID for which to obtain the data from the
\code{currencies} table.}

\item{currency}{The currency code (e.g., \code{"USD"}, \code{"EUR"}, etc.).}

\item{param}{The parameter from which information will be obtained for the
specified \code{locale}.}
}
\value{
A length one character vector.
}
\description{
The \code{currencies} table contains information of currency codes and localized
display names and symbols across 571 locales. The \code{cldr_currency()} function
allows one to extract a single parameter value from the table by supplying
the \code{locale}, the currency code, and one of the following parameter names:
\itemize{
\item \code{"sym"}: corresponds to the \code{currency_symbol} column in \code{currencies}
\item \code{"sym_narrow"}: is the \code{currency_symbol_narrow} column in \code{currencies}
\item \code{"displ_name"}: is \code{currency_display_name}
\item \code{"displ_ct_1"}: is \code{currency_display_name_count_1}
\item \code{"displ_ct_oth"}: is \code{currency_display_name_count_other}
}
}
