% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{currencies}
\alias{currencies}
\title{A table with localized currency attributes and descriptors}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 173013 rows and 7 columns.
}
\usage{
currencies
}
\description{
The \code{currencies} table contains localized data for number-related entities
across 571 locales. This table has 173,013 rows, one per
distinct combination of locale and currency (\code{currency_code}), and the
following 7 columns:
\itemize{
\item \code{locale} (\code{character})
\item \code{currency_code} (\code{character})
\item \code{currency_display_name} (\code{character})
\item \code{currency_symbol} (\code{character})
\item \code{currency_symbol_narrow} (\code{character})
\item \code{currency_display_name_count_1} (\code{character})
\item \code{currency_display_name_count_other} (\code{character})
}
}
\keyword{datasets}
