% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dates}
\alias{dates}
\title{A table with localized date attributes and descriptors}
\format{
A tibble with 571 rows and 29 variables:
\describe{
\item{locale}{The locale name.}
\item{months_format_abbrev}{months_format_abbrev}
\item{months_format_narrow}{months_format_narrow}
\item{months_format_wide}{months_format_wide}
\item{days_standalone_narrow}{days_standalone_narrow}
\item{days_standalone_short}{days_standalone_short}
\item{days_standalone_wide}{days_standalone_wide}
\item{quarters_format_abbrev}{quarters_format_abbrev}
\item{quarters_format_narrow}{quarters_format_narrow}
\item{quarters_format_wide}{quarters_format_wide}
\item{quarters_standalone_abbrev}{quarters_standalone_abbrev}
\item{quarters_standalone_narrow}{quarters_standalone_narrow}
\item{quarters_standalone_wide}{quarters_standalone_wide}
\item{dayperiods_format_abbrev}{dayperiods_format_abbrev}
\item{dayperiods_format_narrow}{dayperiods_format_narrow}
\item{dayperiods_format_wide}{dayperiods_format_wide}
\item{dayperiods_standalone_abbrev}{dayperiods_standalone_abbrev}
\item{dayperiods_standalone_narrow}{dayperiods_standalone_narrow}
\item{dayperiods_standalone_wide}{dayperiods_standalone_wide}
\item{eras_abbrev}{eras_abbrev}
\item{eras_names}{eras_names}
\item{eras_narrow}{eras_narrow}
\item{date_formats}{date_formats}
\item{date_skeletons}{date_skeletons}
\item{time_formats}{time_formats}
\item{time_skeletons}{time_skeletons}
\item{date_time_available_formats}{date_time_available_formats}
\item{date_time_append_items}{date_time_append_items}
\item{date_time_interval_formats}{date_time_interval_formats}
}
}
\usage{
dates
}
\description{
The \code{dates} table contains localized data for constructing dates and times
across 571 locales.
}
\keyword{datasets}
