% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tz_exemplar}
\alias{tz_exemplar}
\title{A table with localized names for all time zone exemplar cities}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 574 rows and 443 columns.
}
\usage{
tz_exemplar
}
\description{
The \code{tz_exemplar} table contains localized names for all exemplar cities used
in time zone names. There are 574 rows and a column for
each exemplar city name (comprising 442 columns; the
\code{locale} column is first). To have syntactical column names, all slashes in
exemplar city names are instead represented with period characters (e.g.,
\code{Indiana/Vincennes} is \code{Indiana.Vincennes}). Some exemplar cities are not
actually cities and these are: \code{UTC.long.standard}
(\emph{en}: \code{"Coordinated Universal Time"}), \code{UTC.short.standard} (\emph{en}: \code{"UTC"}),
and \code{Unknown} (\emph{en}: \code{"Unknown City"}).
}
\keyword{datasets}
